frontend/requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
WTForms==3.0.1
python-dotenv==1.0.0
Flask-SocketIO==5.3.6
gevent-websocket==0.10.1
requests==2.31.0
gevent==24.2.1
gevent-websocket==0.10.1
urllib3==2.0.7
python-socketio==5.10.0
python-engineio==4.8.0
pytest==7.4.2
uvicorn==0.22.0
